+++ deltafi-ui/src/components/DialogTemplate.vue
@@ -21,7 +21,7 @@
     <span @click="showDialog()">
       <slot />
     </span>
-    <Dialog v-model:visible="dialogVisible" :header="$attrs['header']" position="top" :style="{ width: dialogSize }" :maximizable="true" :modal="true" :dismissable-mask="true" :draggable="false">
+    <Dialog id="dialogTemplate" v-model:visible="dialogVisible" :header="$attrs['header']" position="top" :style="{ width: dialogSize }" :maximizable="true" :modal="true" :dismissable-mask="true" :draggable="false">
       <Component :is="loadComponent" v-bind="$attrs" :close-dialog-command="closeDialogCommand" />
     </Dialog>
   </div>